+++ core/src/com/bot4s/telegram/models/Contact.scala
@@ -12,6 +12,6 @@ case class Contact(
                     phoneNumber : String,
                     firstName   : String,
                     lastName    : Option[String] = None,
-                    userId      : Option[Int] = None,
+                    userId      : Option[Long] = None,
                     vcard       : Option[String] = None
                   )
